refactor(map): extract file-backed overlay mechanics from squeeze zones
The squeeze-zones overlay owns the generic file-picker, all-or-nothing
parse, localStorage persistence, and self-disable flow. The upcoming
cue-events overlay (#231) needs the identical flow, so lift it into
createFileBackedOverlay and keep squeeze-zones as parse/render config.
No behavior change — toast strings and load semantics are preserved.
